requirements.txt
streamlit>=1.26.0
PyPDF2>=3.0.0
PyMuPDF>=1.22.5
google-generativeai>=0.3.0
python-dotenv>=1.0.0
pandas>=2.0.0